fix build

+++ libs/math/safemath.go
@@ -42,7 +42,7 @@ func SafeConvertInt32(a int64) int32 {
 
 // SafeConvertUint32 takes a int and checks if it overflows
 // If there is an overflow this will panic
-func SafeConvertUint32(a int64) uint32 {
+func SafeConvertUint32(a int) uint32 {
 	if a > math.MaxUint32 {
 		panic(ErrOverflowInt32)
 	} else if a < 0 {
